requirements.txt
sentence-transformers
langchain
flask
pypdf
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain-huggingface
langchain-groq
langchain_community
langchain_openai
langchain_experimental
-e .